webscraper/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
ConfigArgParse==1.7.1
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.16.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
locust==2.34.0
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.1
packaging==24.2
parsel==1.9.1
Protego==0.3.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==27.0.0
queuelib==1.7.0
redis==6.2.0
requests==2.32.3
requests-file==2.1.0
s3transfer==0.10.4
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.3
w3lib==2.2.1
Werkzeug==3.1.3
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
